auth/requirements.txt
fastapi[standard]
peewee
psycopg2-binary
pyjwt
requests